Cleaned up javadoc warnings (#4)
* moves several gradle properties to env vars

* - switches signing property retrieval method from findProperty
to getProperty as to trigger throws on missing properties
- corrects issues with mis matched env variable name for signing key
  password

* switches back to findProperty after identifying mismatches
and guards sign statement with null check on properties

* Adds null check on signing variables in build.gradle and log statement
indicating when the signing process is continueing

* Adds initial javadoc comment blocks to clean up warnings

+++ lib/src/main/java/ninja/codingsolutions/solaredgeapiclient/models/PowerData.java
@@ -4,9 +4,12 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.extern.jackson.Jacksonized;
 
+/**
+ * Power data in kWh
+ */
 @Data
 @Builder
 @Jacksonized
 public class PowerData {
-    private int power;
+    private double power;
 }
